+++ tasks/help_me_carry/requirements.in
@@ -1,0 +1,1 @@
+torch
