--- tasks/restaurant/requirements.txt
@@ -1,1 +1,0 @@
-jellyfish==1.1.0          # via -r requirements.in
